+++ packages/atomic/src/components/atomic-date-facet/atomic-date-facet.tsx
@@ -17,7 +17,7 @@ import {
 @Component({
   tag: 'atomic-date-facet',
   styleUrl: 'atomic-date-facet.pcss',
-  shadow: false,
+  shadow: true,
 })
 export class AtomicDateFacet implements InitializableComponent {
   @InitializeBindings() public bindings!: Bindings;
@@ -26,6 +26,7 @@ export class AtomicDateFacet implements InitializableComponent {
   @BindStateToController('facet', {subscribeOnConnectedCallback: true})
   @State()
   private facetState!: DateFacetState;
+  @State() public error!: Error;
 
   @Prop({mutable: true, reflect: true}) public facetId = '';
   @Prop() public field = '';