+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -33,7 +33,7 @@ export type InitializationOptions = Pick<
 
 @Component({
   tag: 'atomic-search-interface',
-  shadow: false,
+  shadow: true,
   assetsDirs: ['lang'],
 })
 export class AtomicSearchInterface {