+++ packages/samples/headless-react/src/App.tsx
@@ -1,0 +1,26 @@
+import logo from './logo.svg';
+import './App.css';
+import {SearchBox} from './components/search-box/search-box.class';
+import {SearchBox as SearchBoxFn} from './components/search-box/search-box.fn';
+import {buildSearchBox, SearchBoxOptions} from '@coveo/headless';
+import {engine} from './engine';
+import {Section} from './layout/section';
+
+const options: SearchBoxOptions = {numberOfSuggestions: 8};
+const searchBox = buildSearchBox(engine, {options});
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <Section title="search-box">
+          <SearchBox />
+          <SearchBoxFn controller={searchBox} />
+        </Section>
+      </header>
+    </div>
+  );
+}
+
+export default App;