+++ packages/samples/headless-react/src/engine.ts
@@ -1,0 +1,6 @@
+import {HeadlessEngine, searchAppReducers} from '@coveo/headless';
+
+export const engine = new HeadlessEngine({
+  configuration: HeadlessEngine.getSampleConfiguration(),
+  reducers: searchAppReducers,
+});